Lots of QOL stuff. Image captures and saves faster (also indicates to user that confirm is not quite ready right away)
Text fields in EditEntry capitalize properly

Capture and confirm buttons on MunchCam look nice

Discovered and fixed bug where user pressed back while editing an existing entry, causing the image to be lost

Tapping any part of the search view in MainActivity puts it in focus for typing

+++ app/src/main/java/seniorproj/munchbox/EditEntry.java
@@ -196,14 +196,17 @@ public class EditEntry extends Activity {
         finish();
         Intent backToList = new Intent(EditEntry.this, MainActivity.class);
         //Delete image for cancelled entry
-        File toDelete = new File(imgPath);
-        if (toDelete.exists()) {
-            if (toDelete.delete()) {
-                System.out.println("File deleted");
-            } else {
-                System.out.println("File not deleted");
+        if (id == -1) {
+            File toDelete = new File(imgPath);
+            if (toDelete.exists()) {
+                if (toDelete.delete()) {
+                    System.out.println("File deleted");
+                } else {
+                    System.out.println("File not deleted");
+                }
             }
         }
         startActivity(backToList);
     }
+    //TODO I'd like the Save button to disappear when the keyboard is up. Related: tap outside keyboard to close it?
 }
